Agregado crud: tipo de contrato del trabajador, tipo de inactividad del trabajador, tipo del sector al que pertenece una organización en la cual el trabajador ya ha laborado. Información Laboral de los trabajadores en desarrollo

+++ modules/Payroll/Database/Migrations/2019_04_24_144029_create_payroll_childrens_table.php
@@ -13,21 +13,23 @@ class CreatePayrollChildrensTable extends Migration
      */
     public function up()
     {
-        Schema::create('payroll_childrens', function (Blueprint $table) {
-            $table->increments('id');
-            $table->string('first_name', 100)->comment('Nombre del hijo del trabajador');
-            $table->string('last_name', 100)->comment('Apellido del hijo del trabajador');
-            $table->string('id_number', 12)->nullable()->comment('Cédula del hijo del trabajador');
-            $table->date('birthdate')->comment('Fecha de nacimiento del hijo del trabajador');
+        if (!Schema::hasTable('payroll_childrens')) {
+            Schema::create('payroll_childrens', function (Blueprint $table) {
+                $table->increments('id');
+                $table->string('first_name', 100)->comment('Nombre del hijo del trabajador');
+                $table->string('last_name', 100)->comment('Apellido del hijo del trabajador');
+                $table->string('id_number', 12)->nullable()->comment('Cédula del hijo del trabajador');
+                $table->date('birthdate')->comment('Fecha de nacimiento del hijo del trabajador');
 
-            $table->integer('payroll_socioeconomic_information_id')->unsigned()
-                  ->comment('identificador de la información socioeconómica que pertenece al hijo');
-            $table->foreign('payroll_socioeconomic_information_id')->references('id')->on('payroll_socioeconomic_informations')
-                  ->onDelete('restrict')->onUpdate('cascade');
+                $table->integer('payroll_socioeconomic_information_id')->unsigned()
+                      ->comment('identificador de la información socioeconómica que pertenece al hijo');
+                $table->foreign('payroll_socioeconomic_information_id')->references('id')->on('payroll_socioeconomic_informations')
+                      ->onDelete('restrict')->onUpdate('cascade');
 
-            $table->timestamps();
-            $table->softDeletes()->comment('Fecha y hora en la que el registro fue eliminado');
-        });
+                $table->timestamps();
+                $table->softDeletes()->comment('Fecha y hora en la que el registro fue eliminado');
+            });
+        }
     }
 
     /**
